feat: add configured DateRangePicker, add ModularFormsDateRangePicker, update types for StepProps
