+++ client/src/feat/my-passport-form/steps/personal-details/index.tsx
@@ -1,11 +1,5 @@
-import {
-	parseDate,
-	type DatePickerValueChangeDetails,
-} from "@ark-ui/solid/date-picker";
-import { getValue, setValue, type FieldEvent } from "@modular-forms/solid";
+import { getValue, type FieldEvent } from "@modular-forms/solid";
 import { useI18n } from "core/context/i18n";
-import { parse } from "date-fns";
-import { invariant, partial } from "es-toolkit";
 import type { resources } from "feat/my-passport-form/resources/i18n/en-US";
 import {
 	type MyPassportForm,
@@ -14,8 +8,7 @@ import {
 import { AutocorrectTextField } from "feat/my-passport-form/ui/autocorrect-text-field";
 import { InputGroup } from "feat/my-passport-form/ui/input-group";
 import { NextRow } from "feat/my-passport-form/ui/next-row";
-import { Index, Show, type Component } from "solid-js";
-import { Portal } from "solid-js/web";
+import { Show, type Component } from "solid-js";
 import {
 	Combobox,
 	ComboboxClearSelection,
@@ -25,26 +18,7 @@ import {
 	ComboboxTrigger,
 } from "ui/combobox";
 import { ModularFormsCombobox } from "ui/combobox/modular-forms-combobox";
-import {
-	DatePicker,
-	DatePickerContent,
-	DatePickerContext,
-	DatePickerControl,
-	DatePickerInput,
-	DatePickerPositioner,
-	DatePickerRangeText,
-	DatePickerTable,
-	DatePickerTableBody,
-	DatePickerTableCell,
-	DatePickerTableCellTrigger,
-	DatePickerTableHead,
-	DatePickerTableHeader,
-	DatePickerTableRow,
-	DatePickerTrigger,
-	DatePickerView,
-	DatePickerViewControl,
-	DatePickerViewTrigger,
-} from "ui/date-picker";
+import { ModularFormsDateRangePicker } from "ui/date-picker/modular-forms-date-range-picker";
 import { Label } from "ui/label";
 import {
 	SelectClearSelection,
@@ -379,7 +353,7 @@ export const PersonalDetails: Component<StepProps> = props => {
 												getValue(
 													props.form,
 													"personalDetails.height",
-												),
+												) as number,
 											)}>
 											{t(
 												"form.personalDetails.height.descriptionCentimetres",
@@ -391,7 +365,7 @@ export const PersonalDetails: Component<StepProps> = props => {
 												getValue(
 													props.form,
 													"personalDetails.height",
-												),
+												) as number,
 											)}>
 											{t(
 												"form.personalDetails.height.descriptionMetres",
@@ -417,14 +391,12 @@ export const PersonalDetails: Component<StepProps> = props => {
 										)}
 									</Label>
 
-									<InputDate
+									<ModularFormsDateRangePicker
+										form={props.form}
+										{...field}
 										{...fieldProps}
-										value={field.value}
-										onChange={partial(
-											setValue,
-											props.form,
-											"personalDetails.dateOfBirth",
-										)}
+										/// @ts-expect-error: expects a div not an input
+										ref={fieldProps.ref}
 										placeholder={t(
 											"form.personalDetails.dateOfBirth.placeholder",
 										)}
@@ -553,197 +525,6 @@ export const PersonalDetails: Component<StepProps> = props => {
 				)}
 			</props.Field>
 		</>
-	);
-};
-
-// TODO
-const InputDate = (props: any) => {
-	const onChange = (details?: DatePickerValueChangeDetails) => {
-		if (!details?.valueAsString.length) {
-			props.onChange(null);
-
-			return;
-		}
-
-		const selectedDate = details.valueAsString.at(0);
-
-		invariant(selectedDate, "Selected date is not specified");
-
-		const parsedDate = parse(selectedDate ?? "", "dd/MM/yyyy", new Date());
-
-		if (!import.meta.env.PROD) {
-			console.debug("parsed date", parsedDate);
-		}
-
-		props.onChange?.(parse(selectedDate ?? "", "dd/MM/yyyy", new Date()));
-	};
-
-	const getValue = () => {
-		if (!props.value || !(props.value instanceof Date)) {
-			return;
-		}
-
-		return [parseDate(props.value)];
-	};
-
-	return (
-		<DatePicker value={getValue()} onValueChange={onChange}>
-			<DatePickerControl class="w-full">
-				<DatePickerInput placeholder={props.placeholder} />
-				<DatePickerTrigger />
-			</DatePickerControl>
-			<Portal>
-				<DatePickerPositioner>
-					<DatePickerContent>
-						<DatePickerView view="day">
-							<DatePickerContext>
-								{context => (
-									<>
-										<DatePickerViewControl>
-											<DatePickerViewTrigger>
-												<DatePickerRangeText />
-											</DatePickerViewTrigger>
-										</DatePickerViewControl>
-										<DatePickerTable>
-											<DatePickerTableHead>
-												<DatePickerTableRow>
-													<Index
-														each={
-															context().weekDays
-														}>
-														{weekDay => (
-															<DatePickerTableHeader>
-																{
-																	weekDay()
-																		.short
-																}
-															</DatePickerTableHeader>
-														)}
-													</Index>
-												</DatePickerTableRow>
-											</DatePickerTableHead>
-											<DatePickerTableBody>
-												<Index each={context().weeks}>
-													{week => (
-														<DatePickerTableRow>
-															<Index
-																each={week()}>
-																{day => (
-																	<DatePickerTableCell
-																		value={day()}>
-																		<DatePickerTableCellTrigger>
-																			{
-																				day()
-																					.day
-																			}
-																		</DatePickerTableCellTrigger>
-																	</DatePickerTableCell>
-																)}
-															</Index>
-														</DatePickerTableRow>
-													)}
-												</Index>
-											</DatePickerTableBody>
-										</DatePickerTable>
-									</>
-								)}
-							</DatePickerContext>
-						</DatePickerView>
-						<DatePickerView view="month">
-							<DatePickerContext>
-								{context => (
-									<>
-										<DatePickerViewControl>
-											<DatePickerViewTrigger>
-												<DatePickerRangeText />
-											</DatePickerViewTrigger>
-										</DatePickerViewControl>
-										<DatePickerTable>
-											<DatePickerTableBody>
-												<Index
-													each={context().getMonthsGrid(
-														{
-															columns: 4,
-															format: "short",
-														},
-													)}>
-													{months => (
-														<DatePickerTableRow>
-															<Index
-																each={months()}>
-																{month => (
-																	<DatePickerTableCell
-																		value={
-																			month()
-																				.value
-																		}>
-																		<DatePickerTableCellTrigger>
-																			{
-																				month()
-																					.label
-																			}
-																		</DatePickerTableCellTrigger>
-																	</DatePickerTableCell>
-																)}
-															</Index>
-														</DatePickerTableRow>
-													)}
-												</Index>
-											</DatePickerTableBody>
-										</DatePickerTable>
-									</>
-								)}
-							</DatePickerContext>
-						</DatePickerView>
-						<DatePickerView view="year">
-							<DatePickerContext>
-								{context => (
-									<>
-										<DatePickerViewControl>
-											<DatePickerViewTrigger>
-												<DatePickerRangeText />
-											</DatePickerViewTrigger>
-										</DatePickerViewControl>
-										<DatePickerTable>
-											<DatePickerTableBody>
-												<Index
-													each={context().getYearsGrid(
-														{
-															columns: 4,
-														},
-													)}>
-													{years => (
-														<DatePickerTableRow>
-															<Index
-																each={years()}>
-																{year => (
-																	<DatePickerTableCell
-																		value={
-																			year()
-																				.value
-																		}>
-																		<DatePickerTableCellTrigger>
-																			{
-																				year()
-																					.label
-																			}
-																		</DatePickerTableCellTrigger>
-																	</DatePickerTableCell>
-																)}
-															</Index>
-														</DatePickerTableRow>
-													)}
-												</Index>
-											</DatePickerTableBody>
-										</DatePickerTable>
-									</>
-								)}
-							</DatePickerContext>
-						</DatePickerView>
-					</DatePickerContent>
-				</DatePickerPositioner>
-			</Portal>
-		</DatePicker>
 	);
 };
